+++ packages/gatsby/package.json
@@ -75,6 +75,7 @@
     "invariant": "^2.2.4",
     "is-relative": "^1.0.0",
     "is-relative-url": "^2.0.0",
+    "jest-worker": "^23.2.0",
     "joi": "12.x.x",
     "json-loader": "^0.5.7",
     "json-stringify-safe": "^5.0.1",
@@ -98,6 +99,7 @@
     "opn": "^5.3.0",
     "parse-filepath": "^1.0.1",
     "path-exists": "^3.0.0",
+    "physical-cpu-count": "^2.0.0",
     "postcss-flexbugs-fixes": "^3.0.0",
     "postcss-loader": "^2.1.3",
     "raw-loader": "^0.5.1",
