+++ packages/gatsby/src/commands/build-html.js
@@ -6,7 +6,7 @@ const debug = require(`debug`)(`gatsby:html`)
 const webpackConfig = require(`../utils/webpack.config`)
 const { store } = require(`../redux`)
 const { createErrorFromString } = require(`gatsby-cli/lib/reporter/errors`)
-const renderHTML = require(`../utils/html-renderer`)
+const renderHTMLQueue = require(`../utils/html-renderer-queue`)
 
 module.exports = async (program: any, activity: any) => {
   const { directory } = program
@@ -41,7 +41,7 @@ module.exports = async (program: any, activity: any) => {
         )
       }
 
-      return renderHTML(require(outputFile), pages, activity)
+      return renderHTMLQueue(outputFile, pages, activity)
         .then(() => {
           // Remove the temp JS bundle file built for the static-site-generator-plugin
           try {
